osa 4 d ready

+++ osa4/blogilista/controllers/blogs.js
@@ -1,29 +1,54 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const {request, response} = require("express");
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
-blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs)
-  })
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+      .find({}).populate('user', {username: 1, name: 1})
+  response.json(blogs)
 })
 
 
-blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
+  const body = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
 
   try {
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
     const savedBlog = await blog.save()
+    user.blogs =user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Only creator can delete a blog' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  response.status(204).json({confirmation: 'a blog deleted'}).end()
   } catch (error) {
     next(error)
   }
